* Source/NSPrinter.m: * Source/NSPrintOperation.m: * Source/NSPageLayout.m: * Source/NSPrintInfo.m: * Headers/AppKit/NSPrintInfo.h: Migrate printing code to CGFloat or double. * Source/NSTextView.m: Rewrite -adjustPageHeightNew:top:bottom:limit:
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@34288 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSPrintInfo.h
@@ -76,20 +76,20 @@ typedef enum _NSPrintingPaginationMode {
 // Managing the Printing Rectangle 
 //
 + (NSSize)sizeForPaperName:(NSString *)name;
-- (float)bottomMargin;
-- (float)leftMargin;
+- (CGFloat)bottomMargin;
+- (CGFloat)leftMargin;
 - (NSPrintingOrientation)orientation;
 - (NSString *)paperName;
 - (NSSize)paperSize;
-- (float)rightMargin;
-- (void)setBottomMargin:(float)value;
-- (void)setLeftMargin:(float)value;
+- (CGFloat)rightMargin;
+- (void)setBottomMargin:(CGFloat)value;
+- (void)setLeftMargin:(CGFloat)value;
 - (void)setOrientation:(NSPrintingOrientation)mode;
 - (void)setPaperName:(NSString *)name;
 - (void)setPaperSize:(NSSize)size;
-- (void)setRightMargin:(float)value;
-- (void)setTopMargin:(float)value;
-- (float)topMargin;
+- (void)setRightMargin:(CGFloat)value;
+- (void)setTopMargin:(CGFloat)value;
+- (CGFloat)topMargin;
 #if OS_API_VERSION(MAC_OS_X_VERSION_10_2, GS_API_LATEST)
 - (NSRect)imageablePageBounds;
 #endif
